Add DoubleFuck preprocessor
This commit adds syntax checking preprocessing function for DoubleFuck.
There are currently no special preprocessing functions, only noop can be
used with DoubleFuck.

Preprocessing bracket matching function was moved from BF to utils and
was generalized to look for any characters specified. It only looks on
one pair of characters at a time. The noop preprocessing function was
also moved to utils, with no changes to its behavior.

With these changes to checking and preprocessing functions, exceptions
indicating syntax errors in the interpreted program were moved to utils
and generalized.