Create better check for brackets
The new brackets check is an upgraded version of the one added before.
It can check for multiple bracket types at a time, keeping track of
whether they are closed in the reverse order to how they were open.
The following sequence is one of the examples that is only detected as
error by the new function and not the old one: `( [ ) ]`